Fix size of TARGET to 32 bit
Multiplayer depends on this. Requires a big pile of pointer nastiness and breaks saves (where long is 64bit)

+++ CODE/defines.h
@@ -567,7 +567,7 @@ typedef int		WAYPOINT;
 **	and an index value of 0, the target value returned is identical with
 **	TARGET_NONE. This is by design and is necessary.
 */
-typedef long	TARGET;
+typedef int32_t	TARGET;
 
 #define	TARGET_MANTISSA			24				// Bits of value precision.
 #define	TARGET_EXPONENT			8
